"Added Main Page"
Close #3. This commit adds a beautiful front and center page to our
website, describing all sorts of info on the goals of the organization.
With detailed descriptions including the goals of the org and the
procedure, this page aims to inform the veiwer what we do with the money
and the goals of our organization.